requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.1.0
psycopg2-binary==2.9.6
pip==22.3.1
pytz==2023.3
setuptools==65.5.1
sqlparse==0.4.4
tzdata==2023.3
wheel==0.38.4
drf-spectacular==0.26.3
markdown==3.4.3
gunicorn==20.1.0
